+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.10.0-dev3
 
+* Made a number of `State` classes private.
 * Added: `popUntil` to pop multiple times with a predicate.
 * Fixed: when navigating to relative route, use the current context's path as
   the base, instead of the router's current path.