+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # 0.10.0-dev3
 
+* Added: `history` property on `Routemaster` for chronological history navigation,
+  for example `Routemaster.of(context).history.back()`.
+
+# 0.10.0-dev3
+
 * Added: `popUntil` to pop multiple times with a predicate.
-* Fixed: when navigating to relative route, use the current context's path as
+* Fixed: when navigating to a relative route, use the current context's path as
   the base, instead of the router's current path.
 
 # 0.10.0-dev2